requirements.txt
streamlit
torch
torchvision
opencv-python
numpy
Pillow
dlib==19.22.99
